+++ ccrstore.go
@@ -1,5 +1,0 @@
-package headroom
-
-import "github.com/superops-team/headroom-go/internal/types"
-
-type CCRStore = types.CCRStore